:building_construction: Extract Suggestion to Independent Module
